+++ client/src/jsonrpc.rs
@@ -45,7 +45,24 @@ pub trait Rpc {
         account: Option<String>,
         id: String,
         recipient: Option<String>,
-        group_id: Option<String>,
+        #[allow(non_snake_case)] groupId: Option<String>,
+    ) -> Result<Value, ErrorObjectOwned>;
+
+    #[method(name = "getAvatar", param_kind = map)]
+    fn get_avatar(
+        &self,
+        account: Option<String>,
+        contact: Option<String>,
+        profile: Option<String>,
+        #[allow(non_snake_case)] groupId: Option<String>,
+    ) -> Result<Value, ErrorObjectOwned>;
+
+    #[method(name = "getSticker", param_kind = map)]
+    fn get_sticker(
+        &self,
+        account: Option<String>,
+        #[allow(non_snake_case)] packId: String,
+        #[allow(non_snake_case)] stickerId: u32,
     ) -> Result<Value, ErrorObjectOwned>;
 
     #[method(name = "getUserStatus", param_kind = map)]